move regularized_cov in cov.py
