+++ mne/decoding/csp.py
@@ -6,83 +6,9 @@
 
 import numpy as np
 from scipy import linalg
-import six
-from distutils.version import LooseVersion
 
 from .mixin import TransformerMixin
-
-
-def _regularized_covariance(data, reg=None):
-    """Compute a regularized covariance from data.
-
-    Parameters
-    ----------
-    data : ndarray, shape (n_channels, n_times)
-        Data for covariance estimation.
-    reg : float | str | None (default None)
-        If not None, allow regularization for covariance estimation
-        if float, shrinkage covariance is used (0 <= shrinkage <= 1).
-        if str, optimal shrinkage using Ledoit-Wolf Shrinkage ('lws') or
-        Oracle Approximating Shrinkage ('oas').
-
-    Returns
-    -------
-    cov : ndarray, shape (n_channels, n_channels)
-        The covariance matrix.
-    """
-    if reg is None:
-        # compute empirical covariance
-        cov = np.cov(data)
-    else:
-        no_sklearn_err = ('the scikit-learn package is missing and '
-                          'required for covariance regularization.')
-        # use sklearn covariance estimators
-        if isinstance(reg, float):
-            if (reg < 0) or (reg > 1):
-                raise ValueError('0 <= shrinkage <= 1 for '
-                                 'covariance regularization.')
-            try:
-                import sklearn
-                sklearn_version = LooseVersion(sklearn.__version__)
-                from sklearn.covariance import ShrunkCovariance
-            except ImportError:
-                raise Exception(no_sklearn_err)
-            if sklearn_version < '0.12':
-                skl_cov = ShrunkCovariance(shrinkage=reg,
-                                           store_precision=False)
-            else:
-                # init sklearn.covariance.ShrunkCovariance estimator
-                skl_cov = ShrunkCovariance(shrinkage=reg,
-                                           store_precision=False,
-                                           assume_centered=True)
-        elif isinstance(reg, six.string_types):
-            if reg == 'lws':
-                try:
-                    from sklearn.covariance import LedoitWolf
-                except ImportError:
-                    raise Exception(no_sklearn_err)
-                # init sklearn.covariance.LedoitWolf estimator
-                skl_cov = LedoitWolf(store_precision=False,
-                                     assume_centered=True)
-            elif reg == 'oas':
-                try:
-                    from sklearn.covariance import OAS
-                except ImportError:
-                    raise Exception(no_sklearn_err)
-                # init sklearn.covariance.OAS estimator
-                skl_cov = OAS(store_precision=False,
-                              assume_centered=True)
-            else:
-                raise ValueError("regularization parameter should be "
-                                 "'lwf' or 'oas'")
-        else:
-            raise ValueError("regularization parameter should be "
-                             "of type str or int (got %s)." % type(reg))
-
-        # compute regularized covariance using sklearn
-        cov = skl_cov.fit(data.T).covariance_
-
-    return cov
+from ..cov import _regularized_covariance
 
 
 class CSP(TransformerMixin):
